Fix compile warnings.

+++ Libs/System Services/Utilities/SSNetworkInfo.m
@@ -674,8 +674,11 @@
         NSUInteger BA = ~NM | IP;
         
         // Make a string for the address
-        NSString *BroadcastAddress = [NSString stringWithFormat:@"%lu.%lu.%lu.%lu", (BA & 0xFF000000) >> 24,
-                                      (BA & 0x00FF0000) >> 16, (BA & 0x0000FF00) >> 8, BA & 0x000000FF];
+        NSString *BroadcastAddress = [NSString stringWithFormat:@"%lu.%lu.%lu.%lu",
+                                        (unsigned long) (BA & 0xFF000000) >> 24,
+                                        (unsigned long) (BA & 0x00FF0000) >> 16,
+                                        (unsigned long) (BA & 0x0000FF00) >> 8,
+                                        (unsigned long) BA & 0x000000FF];
         
         // Check to make sure the string is valid
         if (BroadcastAddress == nil || BroadcastAddress.length <= 0) {
@@ -989,8 +992,11 @@
         NSUInteger BA = ~NM | IP;
         
         // Make a string for the address
-        NSString *BroadcastAddress = [NSString stringWithFormat:@"%lu.%lu.%lu.%lu", (BA & 0xFF000000) >> 24,
-                                      (BA & 0x00FF0000) >> 16, (BA & 0x0000FF00) >> 8, BA & 0x000000FF];
+        NSString *BroadcastAddress = [NSString stringWithFormat:@"%lu.%lu.%lu.%lu",
+                                        (unsigned long) (BA & 0xFF000000) >> 24,
+                                        (unsigned long) (BA & 0x00FF0000) >> 16,
+                                        (unsigned long) (BA & 0x0000FF00) >> 8,
+                                        (unsigned long) BA & 0x000000FF];
         
         // Check to make sure the string is valid
         if (BroadcastAddress == nil || BroadcastAddress.length <= 0) {
